Gyan_Intent/backend/pyproject.toml
[project]
name = "gyan-intent-backend"
version = "1.0.0"
description = "Intent-Aware Learning Engine Backend"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-multipart>=0.0.20",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.28.0",
    "langgraph>=0.2.60",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.30",
    "chromadb>=1.5.5",
    "sqlalchemy>=2.0.36",
    "asyncpg>=0.30.0",
    "redis>=5.2.0",
    "celery>=5.4.0",
    "manim>=0.18.0",
    "pillow>=11.0.0",
    "numpy>=2.2.0",
    "opencv-python>=4.10.0",
    "websockets>=14.1",
    "python-dotenv>=1.0.0",
    "structlog>=24.4.0",
    "prometheus-client>=0.21.0",
    "youtube-transcript-api>=0.6.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "black>=24.10.0",
    "ruff>=0.8.0",
    "mypy>=1.14.0",
    "pre-commit>=4.0.0",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
